% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aoi.R
\name{aoi}
\alias{aoi}
\title{Create area of interest (aka landscape area)}
\usage{
aoi(property = NA, centroid = NA, size = 2500, expand = T)
}
\arguments{
\item{property}{optional sf polygon object with geometry of property for
which landscape context is being considered. CRS must be defined. If
\code{property} is not used, \code{centroid} must be defined}

\item{centroid}{optional sf point object with geometry of aoi centroid. CRS
must be defined. If \code{centroid} is not defined, \code{property} must be
used.}

\item{size}{minimum size of aoi, in acres. Defaults to 2500 acres based on
Audubon recommendations.}

\item{expand}{should aoi be allowed to expand in size, to accommodate large
properties that would otherwise extend beyond aoi? TRUE or FALSE. Only used
if \code{property} is defined.}
}
\value{
sf polygon object with area of interest geometry. CRS is
EPSG:32145 (NAD83 / Vermont)
}
\description{
Create area of interest (aka landscape area)
}
