% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{centroid}
\alias{centroid}
\title{Create centroid from lat and lon}
\usage{
centroid(lat, lon)
}
\arguments{
\item{lat}{latitude in decimal degrees}

\item{lon}{longitude in decimal degrees}
}
\value{
sf point object with centroid geometry. CRS is ESPG:32145 (NAD83 /
Vermont)
}
\description{
Manually enter latitude and longitude to create an sf centroid object that
can later be used to define an area of interest, for example.
}
