% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_landcover}
\alias{get_landcover}
\title{Landcover data for area of interest}
\usage{
get_landcover(aoi)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
defined}
}
\value{
\code{RasterBrick} object with RGB color image of landcover classes
in aoi
}
\description{
Retrieves 2016, 0.5m resolution Vermont landcover data for area of interest
from the Vermont Center for Geographic Information, and converts it to a
\code{RasterBrick} object
}
