% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ndsm}
\alias{get_ndsm}
\title{nDSM data for area of interest}
\usage{
get_ndsm(aoi)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
defined}
}
\value{
\code{RasterBrick} object with heights above ground (m) in aoi
}
\description{
Retrieves most recent normalized digital surface model for area of interest
from the Vermont Center for Geographic Information, and converts it to a
SpatRaster object
}
