% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_treetype}
\alias{get_treetype}
\title{Hardwood and softwood canopy data for area of interest}
\usage{
get_treetype(aoi)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
defined.}
}
\value{
\code{SpatRaster} object with tree type class values for aoi. Classes
include Coniferous and Deciduous.
}
\description{
Retrieves 2022 Vermont modeled tree type data for area of interest from local
raster data. Requires \code{set_data_path()} to be configured first.
}
\examples{
\dontrun{
# Set data path first
set_data_path("~/vthabitat_data")

# Get canopy type for an area of interest
pt <- centroid(44.393, -72.487)
my_aoi <- aoi(centroid = pt, size = 100)
treetype <- treetype(my_aoi)
terra::plot(treetype)
}
}
