% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_wetlands}
\alias{get_wetlands}
\title{Wetlands data for area of interest}
\usage{
get_wetlands(aoi)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
defined.}
}
\value{
\code{SpatRaster} object with wetland class values for aoi. Classes
include Emergent, Forested, and Scrub/Shrub wetlands.
}
\description{
Retrieves 2022 Vermont modeled wetlands data for area of interest from local
raster data. Requires \code{set_data_path()} to be configured first.
}
\examples{
\dontrun{
# Set data path first
set_data_path("~/vthabitat_data")

# Get wetlands for an area of interest
pt <- centroid(44.393, -72.487)
my_aoi <- aoi(centroid = pt, size = 100)
wetlands <- get_wetlands(my_aoi)
terra::plot(wetlands)
}
}
