% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{has_data_path}
\alias{has_data_path}
\title{Check if data path is configured}
\usage{
has_data_path()
}
\value{
Logical indicating if data path has been set.
}
\description{
Check if data path is configured
}
\keyword{internal}
