% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\name{make_shrub}
\alias{make_shrub}
\title{Polygons of young forest and shrubland}
\usage{
make_shrub(canopy, size = 1, distance = 100)
}
\arguments{
\item{canopy}{\code{SpatRaster} object showing canopy attributes for area of
interest, made with \code{make_canopy}. Currently CRS must be in meters, or
areas will fail silently!}

\item{size}{minimum size, in acres, a shrubby area must occupy for inclusion}

\item{distance}{maximum distance, in feet, two shrub polygons can be from
one-another and still count as a single shrubby area}
}
\description{
Polygons of young forest and shrubland
}
