% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_canopy}
\alias{map_canopy}
\title{Tree canopy map}
\usage{
map_canopy(canopy, prop = NA)
}
\arguments{
\item{canopy}{categorical raster of tree canopy}

\item{prop}{optional polygon layer of property}
}
\value{
Returns map of area of interest, showing its tree canopy types and,
optionally, the property boundaries.
}
\description{
Tree canopy map
}
