% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_context}
\alias{map_context}
\title{Landscape context map}
\usage{
map_context(clr, landcover, prop = NA)
}
\arguments{
\item{clr}{rgb raster of color orthophoto for area of interest}

\item{landcover}{rgb raster of landover for area of interest}

\item{prop}{optional polygon layer of property}
}
\value{
Returns map of area of interest, showing its landcover overlain on an
aerial photo and, optionally, the property boundaries.
}
\description{
Landscape context map
}
\examples{
\dontrun{
# Set data path first
set_data_path("~/vthabitat_data")

# Get data for area of interest
pt <- centroid(44.393, -72.487)
my_aoi <- aoi(centroid = pt, size = 100)
clr <- get_clr(my_aoi)
lc <- get_landcover(my_aoi)

# Create landscape context map
map_context(clr, lc)
}
}
