% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_shrub}
\alias{map_shrub}
\title{Shrub/young forest map}
\usage{
map_shrub(clr, landcover, shrub, prop = NA)
}
\arguments{
\item{clr}{rgb raster of color orthophoto for area of interest}

\item{landcover}{rgb raster of landover for area of interest}

\item{shrub}{polygon layer of shrub/young forest areas}

\item{prop}{optional polygon layer of property}
}
\value{
Returns map of area of interest, showing shrub/young forest areas
overlain on an aerial photo and, optionally, the property boundaries.
}
\description{
Shrub/young forest map
}
