% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rest-service.R
\name{rest_service}
\alias{rest_service}
\title{Fetch data from ArcREST service}
\usage{
rest_service(aoi, base_url, query_params)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
EPSG:3857 and must be defined}

\item{base_url}{service url, ending in service type and layer if applicable
(eg. .../ImageServer/0)}

\item{query_params}{named list of query parameters and their values. "format"
parameter must be defined to set file extension. See service's api
reference for available parameters.}
}
\value{
\code{SpatRaster} object of fetched data, w/ CRS EPSG:3857
}
\description{
Fetch data from ArcREST service
}
