% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locale.R
\name{shp2sf}
\alias{shp2sf}
\title{Convert shapefile to sf object}
\usage{
shp2sf(file)
}
\arguments{
\item{file}{path to .shp file. CRS must be ESPG:32145 (NAD83 / Vermont)}
}
\value{
sf object
}
\description{
Convert shapefile to sf object
}
\examples{
\dontrun{
# Read a shapefile
property <- shp2sf("path/to/property.shp")
}
}
