% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_clr}
\alias{get_clr}
\title{Color orthophoto for area of interest}
\usage{
get_clr(aoi)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
defined.}
}
\value{
\code{SpatRaster} object with RGB color orthoimage for aoi
}
\description{
Retrieves most recent color orthoimagery for area of interest from the
Vermont Center for Geographic Information.
}
\examples{
\dontrun{
# Get color orthophoto for an area of interest
pt <- centroid(44.393, -72.487)
my_aoi <- aoi(centroid = pt, size = 100)
clr <- get_clr(my_aoi)
terra::plotRGB(clr)
}
}
