% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_ndsm}
\alias{get_ndsm}
\title{nDSM data for area of interest}
\usage{
get_ndsm(aoi)
}
\arguments{
\item{aoi}{sf polygon object with area of interest geometry. CRS must be
defined.}
}
\value{
\code{SpatRaster} object with heights above ground (m) in aoi
}
\description{
Retrieves most recent normalized digital surface model for area of interest
from the Vermont Center for Geographic Information.
}
\examples{
\dontrun{
# Get nDSM for an area of interest
pt <- centroid(44.393, -72.487)
my_aoi <- aoi(centroid = pt, size = 100)
ndsm <- get_ndsm(my_aoi)
terra::plot(ndsm)
}
}
