% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summaries.R
\name{land_summary}
\alias{land_summary}
\title{Landcover summary}
\usage{
land_summary(landcover)
}
\arguments{
\item{landcover}{\code{RasterBrick} object with RGB color image of VT 2016
landcover classes, from \code{get_landcover}}
}
\value{
data frame of cover types and the percentage of land area covered by
each
}
\description{
Summarizes percentage of land area in each of eight cover types, based on a
raster layer of Vermont's 2016 landcover data for a given area
}
