% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/canopy.R
\name{make_canopy}
\alias{make_canopy}
\title{Raster of forest canopy type}
\usage{
make_canopy(ndsm, landcover)
}
\arguments{
\item{ndsm}{\code{RasterBrick} object with heights above ground (m) for area
of interest, from \code{get_ndsm}}

\item{landcover}{\code{RasterBrick} object with RGB color image of VT 2016
landcover classes for area of interest, from \code{get_landcover}}
}
\value{
\code{SpatRaster} object showing canopy attributes for area of
interest, including: (1) shrub and young forest cover, (2) poletimber, (3)
partial overstories, and (4) closed overstories
}
\description{
Create SpatRaster showing forest canopy attributes in area of interest, based
on nDSM and landcover data
}
