% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_canopy}
\alias{map_canopy}
\title{Tree canopy map}
\usage{
map_canopy(canopy, prop = NA)
}
\arguments{
\item{canopy}{categorical raster of tree canopy}

\item{prop}{optional polygon layer of property}
}
\value{
Returns map of area of interest, showing its tree canopy types and,
optionally, the property boundaries.
}
\description{
Tree canopy map
}
\examples{
\dontrun{
# Set data path first
set_data_path("~/vthabitat_data")

# Get data and create canopy raster
pt <- centroid(44.393, -72.487)
my_aoi <- aoi(centroid = pt, size = 100)
ndsm <- get_ndsm(my_aoi)
lc <- get_landcover(my_aoi)
canopy <- make_canopy(ndsm, lc)

# Create canopy map
map_canopy(canopy)
}
}
