% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maps.R
\name{map_context}
\alias{map_context}
\title{Landscape context map}
\usage{
map_context(clr, landcover, prop = NA)
}
\arguments{
\item{clr}{rgb raster of color orthophoto for area of interest}

\item{landcover}{rgb raster of landover for area of interest}

\item{prop}{optional polygon layer of property}
}
\value{
Returns map of area of interest, showing its landcover overlain on an
aerial photo and, optionally, the property boundaries.
}
\description{
Landscape context map
}
